requirements.txt
qrcode[pil]
django
djangorestframework
pytz
